+++ frontend/src/lib/routes-protected.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom';
-import { RouteProps } from '../interfaces/route-props';
+import { RouteProps } from '../components/interfaces/route-props';
 
 const ProtectedRoute: React.FC<RouteProps> = ({ isAuthenticated, children }) => {
   return !isAuthenticated ? <Navigate to="/login" replace /> : children;
