+++ frontend/src/lib/routes-public.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom';
-import { RouteProps } from '../interfaces/route-props';
+import { RouteProps } from '../components/interfaces/route-props';
 
 const PublicRoute: React.FC<RouteProps> = ({ isAuthenticated, children }) => {
     return isAuthenticated ? <Navigate to="/" /> : children
